Revert "Improvement: Requirements installation in venv (#686)"
This reverts commit 030bdee19b9ab73f480d7812cca4e4efe4f11cd1.
