+++ .gitignore
@@ -5,7 +5,6 @@
 */__pycache__/
 docs/build/
 venv/
-env/
 */__TEMP_DATA__/
 
 # File extensions
